fix(workflows): Create commit with updated package.json
This adds the [`npm`](https://github.com/semantic-release/npm) and [`git`](https://github.com/semantic-release/git) plugins to semantic-release so that on release, it will update the version specified in `package.json` and generate a new commit with the update.

The result should be a release commit that serves as the HEAD of the release tag we create later in the workflow.

This also fixes a bug in the workflow that prevented us from getting the right tarball.