fix: Remove assert package to avoid circular dependencies
The presence of this package appeared to be causing circular dependency issues with `util` and `inherits`, which led to improperly-generated rule code that could not run.
The built-in assertions provided by node are sufficient for what we need.

A couple of our rollup dependencies are also upgraded to avoid other issues.