feat(rule): Add rule to check that `id` is only used for Relay IDs
In Relay, `id` is [treated as a reserved word and must be a globally unique value](https://github.com/facebook/relay/issues/1682#issuecomment-296393416).
That means when an `id` field is present on a type, it must be the type that Relay expects (a string type).

This linter assumes that an `ID` scalar is present in the schema, and checks that if an `id` field is present, it is using that type.